Conflicts:
	src/server/scripts/EasternKingdoms/Gilneas/gilneas.cpp

+++ src/server/game/Server/WorldSession.cpp
@@ -1112,9 +1112,9 @@ void WorldSession::ProcessQueryCallbacks()
     //- HandleStableSwapPet
     if (_stableSwapCallback.IsReady())
     {
-        uint32 param = _stableSwapCallback.GetParam();
+        uint8 param = _stableSwapCallback.GetParam();
         _stableSwapCallback.GetResult(result);
         HandleStableSwapPetCallback(result, param);
         _stableSwapCallback.FreeResult();
-    }
+    } 
 }
